requirements.txt
awscli==1.16.126
boto3==1.9.116
botocore==1.12.116
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
pyasn1==0.4.5
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
